Introduces Plog API
The Personio API does not provide user specific access tokens. With the
Personio API Secret you would have access to ALL employees in Personio.
This is not what we want. The plog CLI should only be able to log work
for a single employee using a specific token that only grants access to
the employees worklog in Personio.

The Plog API solves this. It acts as a proxy in front of the Personio
API, implements additional authentication and authorisation logic and
interacts with the Personio API on behalf the plog client.

The plog/token endpoint is secured by a function key.
Users who have the privilage to create user specific API tokens needed
to execute the plog push command, need to setup the function key for the
plog/token endpoint as an env variable PLOG_TOKEN_FUNCTION_KEY.